requirements.txt
Django==4.2.16
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
PyMySQL==1.1.0
django-filter==24.3
cryptography
python-dotenv
gunicorn
